backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# Validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Utils
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0